Reran correcting issue wher career start date took Mexican league seasons

+++ Types.h
@@ -39,7 +39,7 @@ struct HitterStats {
 };
 
 struct HitterValue {
-    int pa;
+    float pa;
     float war;
     float off;
     float def;
